[#75] Relocate primitive AST nodes
- Moved `Succeed<T>` and `Fail<T>` from `BahmanM.Flow.Ast.Pure` to `BahmanM.Flow.Ast.Primitive` namespace.
- Updated all references to these types across the codebase to reflect the new namespace.
- This change clarifies the nature of these nodes as fundamental building blocks of the flow AST.
