+++ src/BahmanM.Flow/CustomBehaviourStrategy.cs
@@ -1,9 +1,11 @@
+using BahmanM.Flow.Ast.Primitive;
+
 namespace BahmanM.Flow;
 
 internal class CustomBehaviourStrategy(IBehaviour behaviour) : IBehaviourStrategy
 {
-    public IFlow<T> ApplyTo<T>(Ast.Pure.Succeed<T> node) => behaviour.Apply(node);
-    public IFlow<T> ApplyTo<T>(Ast.Pure.Fail<T> node) => behaviour.Apply(node);
+    public IFlow<T> ApplyTo<T>(Succeed<T> node) => behaviour.Apply(node);
+    public IFlow<T> ApplyTo<T>(Fail<T> node) => behaviour.Apply(node);
     public IFlow<T> ApplyTo<T>(Ast.Create.Sync<T> node) => behaviour.Apply(node);
     public IFlow<T> ApplyTo<T>(Ast.Create.Async<T> node) => behaviour.Apply(node);
     public IFlow<T> ApplyTo<T>(Ast.Create.CancellableAsync<T> node) => behaviour.Apply(node);
