+++ src/BahmanM.Flow/Flow.cs
@@ -4,9 +4,9 @@ namespace BahmanM.Flow;
 
 public static class Flow
 {
-    public static IFlow<T> Succeed<T>(T value) => new Ast.Pure.Succeed<T>(value);
+    public static IFlow<T> Succeed<T>(T value) => new Succeed<T>(value);
 
-    public static IFlow<T> Fail<T>(Exception exception) => new Ast.Pure.Fail<T>(exception);
+    public static IFlow<T> Fail<T>(Exception exception) => new Fail<T>(exception);
 
     public static IFlow<T> Create<T>(Func<T> operation) => new Ast.Create.Sync<T>(() => operation());
