+++ src/BahmanM.Flow/TimeoutStrategy.cs
@@ -1,9 +1,11 @@
+using BahmanM.Flow.Ast.Primitive;
+
 namespace BahmanM.Flow;
 
 internal class TimeoutStrategy(TimeSpan duration) : IBehaviourStrategy
 {
-    public IFlow<T> ApplyTo<T>(Ast.Pure.Succeed<T> node) => node;
-    public IFlow<T> ApplyTo<T>(Ast.Pure.Fail<T> node) => node;
+    public IFlow<T> ApplyTo<T>(Succeed<T> node) => node;
+    public IFlow<T> ApplyTo<T>(Fail<T> node) => node;
 
     public IFlow<T> ApplyTo<T>(Ast.Create.Sync<T> node)
     {
